hexapod_autonomous/requirements.txt
rclpy>=7.0.0
sensor_msgs>=4.0.0
geometry_msgs>=4.0.0
nav_msgs>=4.0.0
std_msgs>=4.0.0
cv_bridge>=4.0.0
tf2_ros>=0.25.0
tf2_geometry_msgs>=0.25.0
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
adafruit-circuitpython-servokit>=1.4.0
adafruit-circuitpython-pca9685>=3.4.0
pillow>=10.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
openai>=1.0.0
pyttsx3>=2.90
pyyaml>=6.0
serial>=0.0.97
RPi.GPIO>=0.7.1
yolov5>=7.0.0
torch>=2.0.0
torchvision>=0.15.0
sounddevice>=0.4.6
scipy>=1.11.0
dlib>=19.24.0
face_recognition>=1.3.0
imutils>=0.5.4
